+++ oslo/torch/_C/__init__.py
@@ -4,6 +4,26 @@ from pathlib import Path
 
 import torch
 from torch.utils import cpp_extension
+from oslo.torch.jit._utils import _set_jit_fusion_options
+
+_SOFTMAX_KERNEL = None
+
+
+def get_softmax_kernel():
+    global _SOFTMAX_KERNEL
+
+    try:
+        if _SOFTMAX_KERNEL is None:
+            _set_jit_fusion_options()
+            _SOFTMAX_KERNEL = SoftmaxBinder().bind()
+    except Exception:
+        raise EnvironmentError(
+            "Failed compiling custom CUDA kernels. "
+            "please check your CUDA environment."
+        )
+
+    return _SOFTMAX_KERNEL
+
 
 DEFAULT_TORCH_EXTENSION_PATH = os.path.join(
     os.path.expanduser("~"),
